Gomez: Rest Usuario Rest Datos Personales Rest Clases auxiliares

+++ src/main/java/dao/DatospersonalesFacadeREST.java
@@ -28,7 +28,10 @@ import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
+import javax.ws.rs.core.Context;
+import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
+import security.JWT;
 
 /**
  *
@@ -99,19 +102,29 @@ public class DatospersonalesFacadeREST extends AbstractFacade<Datospersonales> {
         return em;
     }
 
-    // METOOD PARA LISTAR LOS DATOS PERSONALES
-    @GET
+    // METODO PARA LISTAR LOS DATOS PERSONALES
+    @POST
     @Path("listardatospersonales")
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
-    public String listarDatosPersonales() {
+    public String listarDatosPersonales(@Context HttpHeaders headers) {
         Gson g = new Gson();
+        JsonObject response = new JsonObject();
 
-        TypedQuery<Object[]> query = em.createNamedQuery("Datospersonales.listar", Object[].class);
-        List<Object[]> resultList = query.getResultList();
-        List<Map<String, Object>> listaMapas = listarMapaDatosPersonales(resultList);
+        String tokenClient = extraerTokenHeader(headers);
+        String idToken = JWT.verifyToken(tokenClient);
 
-        return g.toJson(listaMapas);
+        if (idToken != null) {
+            TypedQuery<Object[]> query = em.createNamedQuery("Datospersonales.listar", Object[].class);
+            List<Object[]> resultList = query.getResultList();
+            List<Map<String, Object>> listaMapas = listarMapaDatosPersonales(resultList);
+            response.addProperty("resultado", "ok");
+            response.add("datos", g.toJsonTree(listaMapas));
+        } else {
+            response.addProperty("resultado", "error");
+        }
+
+        return g.toJson(response);
     }
 
     // METODO PARA LISTAR LOS DATOS PERSONALES EN MAPAS
@@ -139,31 +152,41 @@ public class DatospersonalesFacadeREST extends AbstractFacade<Datospersonales> {
     @Path("obtenerdatospersonales")
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
-    public String obtenerDatosPersonalesPorId(int idPersona) {
+    public String obtenerDatosPersonalesPorId(int idPersona, @Context HttpHeaders headers) {
         Gson g = new Gson();
+        JsonObject response = new JsonObject();
 
-        TypedQuery<Object[]> tq = em.createNamedQuery("Datospersonales.obtenerDatosPersonalesPorId", Object[].class);
-        tq.setParameter("idPersona", idPersona);
-        Object[] usuario = tq.getSingleResult();
+        String tokenClient = extraerTokenHeader(headers);
+        String idToken = JWT.verifyToken(tokenClient);
 
-        Map<String, Object> map = new HashMap<>();
-        map.put("idPersona", usuario[0]);
-        map.put("denoTipoDocumento", usuario[1]);
-        map.put("docuPersona", usuario[2]);
-        map.put("ruc", usuario[3]);
-        map.put("apPaPersona", usuario[4]);
-        map.put("apMaPersona", usuario[5]);
-        map.put("nombPersona", usuario[6]);
-        map.put("genePersona", usuario[7]);
-        map.put("fechPersona", usuario[8]);
-        map.put("direPersona", usuario[9]);
-        map.put("celuPersona", usuario[10]);
-        map.put("emailPersona", usuario[11]);
-        map.put("denoDistrito", usuario[12]);
-        map.put("denoProvincia", usuario[13]);
-        map.put("denoDepartamento", usuario[14]);
+        if (idToken != null) {
+            TypedQuery<Object[]> tq = em.createNamedQuery("Datospersonales.obtenerDatosPersonalesPorId", Object[].class);
+            tq.setParameter("idPersona", idPersona);
+            Object[] usuario = tq.getSingleResult();
 
-        return g.toJson(map);
+            Map<String, Object> map = new HashMap<>();
+            map.put("idPersona", usuario[0]);
+            map.put("denoTipoDocumento", usuario[1]);
+            map.put("docuPersona", usuario[2]);
+            map.put("ruc", usuario[3]);
+            map.put("apPaPersona", usuario[4]);
+            map.put("apMaPersona", usuario[5]);
+            map.put("nombPersona", usuario[6]);
+            map.put("genePersona", usuario[7]);
+            map.put("fechPersona", usuario[8]);
+            map.put("direPersona", usuario[9]);
+            map.put("celuPersona", usuario[10]);
+            map.put("emailPersona", usuario[11]);
+            map.put("denoDistrito", usuario[12]);
+            map.put("denoProvincia", usuario[13]);
+            map.put("denoDepartamento", usuario[14]);
+            response.addProperty("resultado", "ok");
+            response.add("datos", g.toJsonTree(map));
+        } else {
+            response.addProperty("resultado", "error");
+        }
+
+        return g.toJson(response);
     }
 
     // METODO PARA INSERTAR LOS DATOS PERSONALES
@@ -171,92 +194,38 @@ public class DatospersonalesFacadeREST extends AbstractFacade<Datospersonales> {
     @Path("insertardatospersonales")
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
-    public String insertarDatosPersonales(String data) throws ParseException {
+    public String insertarDatosPersonales(String data, @Context HttpHeaders headers) throws ParseException {
         Gson gson = new Gson();
         JsonObject response = new JsonObject();
         JsonObject request = JsonParser.parseString(data).getAsJsonObject();
 
-        String denoTipoDocumento = request.get("denoTipoDocumento").getAsString();
-        String docuPersona = request.get("docuPersona").getAsString();
-        String ruc = request.get("ruc").getAsString();
-        String nombPersona = request.get("nombPersona").getAsString();
-        String apPaPersona = request.get("apPaPersona").getAsString();
-        String apMaPersona = request.get("apMaPersona").getAsString();
-        char genePersona = request.get("genePersona").getAsString().charAt(0);
-        String fecha = request.get("fechPersona").getAsString();
-        String direPersona = request.get("direPersona").getAsString();
-        String celuPersona = request.get("celuPersona").getAsString();
-        String emailPersona = request.get("emailPersona").getAsString();
-        String denoDistrito = request.get("denoDistrito").getAsString();
+        String tokenClient = extraerTokenHeader(headers);
+        String idToken = JWT.verifyToken(tokenClient);
 
-        SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
-        Date fechPersona = sdf.parse(fecha);
+        if (idToken != null) {
+            String denoTipoDocumento = request.get("denoTipoDocumento").getAsString();
+            String docuPersona = request.get("docuPersona").getAsString();
+            String ruc = request.get("ruc").getAsString();
+            String nombPersona = request.get("nombPersona").getAsString();
+            String apPaPersona = request.get("apPaPersona").getAsString();
+            String apMaPersona = request.get("apMaPersona").getAsString();
+            char genePersona = request.get("genePersona").getAsString().charAt(0);
+            String fecha = request.get("fechPersona").getAsString();
+            String direPersona = request.get("direPersona").getAsString();
+            String celuPersona = request.get("celuPersona").getAsString();
+            String emailPersona = request.get("emailPersona").getAsString();
+            String denoDistrito = request.get("denoDistrito").getAsString();
 
-        Tipodocumento tipoDocumento = new TipodocumentoFacadeREST().obtenerTipoDocumento(denoTipoDocumento);
-        Distrito distrito = new DistritoFacadeREST().obtenerDistrito(denoDistrito);
+            SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
+            Date fechPersona = sdf.parse(fecha);
 
-        try {
-            em.getTransaction().begin();
-            Datospersonales datosPersonales = new Datospersonales();
+            Tipodocumento tipoDocumento = new TipodocumentoFacadeREST().obtenerTipoDocumento(denoTipoDocumento);
+            Distrito distrito = new DistritoFacadeREST().obtenerDistrito(denoDistrito);
 
-            datosPersonales.setDocuPersona(docuPersona);
-            datosPersonales.setRuc(ruc);
-            datosPersonales.setNombPersona(nombPersona);
-            datosPersonales.setApPaPersona(apPaPersona);
-            datosPersonales.setApMaPersona(apMaPersona);
-            datosPersonales.setGenePersona(genePersona);
-            datosPersonales.setFechPersona(fechPersona);
-            datosPersonales.setDirePersona(direPersona);
-            datosPersonales.setCeluPersona(celuPersona);
-            datosPersonales.setEmailPersona(emailPersona);
-            datosPersonales.setIdTipoDocumento(tipoDocumento);
-            datosPersonales.setIdDistrito(distrito);
+            try {
+                em.getTransaction().begin();
+                Datospersonales datosPersonales = new Datospersonales();
 
-            em.persist(datosPersonales);
-            em.getTransaction().commit();
-            response.addProperty("success", Boolean.TRUE);
-        } catch (Exception e) {
-            response.addProperty("success", Boolean.FALSE);
-        }
-
-        return gson.toJson(response);
-    }
-
-    // METODO PARA EDITAR LOS DATOS PERSONALES
-    @POST
-    @Path("editardatospersonales")
-    @Consumes(MediaType.APPLICATION_JSON)
-    @Produces(MediaType.APPLICATION_JSON)
-    public String editarDatosPersonales(String data) throws ParseException {
-        Gson gson = new Gson();
-        JsonObject response = new JsonObject();
-        JsonObject request = JsonParser.parseString(data).getAsJsonObject();
-
-        int idPersona = request.get("idPersona").getAsInt();
-        String denoTipoDocumento = request.get("denoTipoDocumento").getAsString();
-        String docuPersona = request.get("docuPersona").getAsString();
-        String ruc = request.get("ruc").getAsString();
-        String nombPersona = request.get("nombPersona").getAsString();
-        String apPaPersona = request.get("apPaPersona").getAsString();
-        String apMaPersona = request.get("apMaPersona").getAsString();
-        char genePersona = request.get("genePersona").getAsString().charAt(0);
-        String fecha = request.get("fechPersona").getAsString();
-        String direPersona = request.get("direPersona").getAsString();
-        String celuPersona = request.get("celuPersona").getAsString();
-        String emailPersona = request.get("emailPersona").getAsString();
-        String denoDistrito = request.get("denoDistrito").getAsString();
-
-        SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
-        Date fechPersona = sdf.parse(fecha);
-
-        Tipodocumento tipoDocumento = new TipodocumentoFacadeREST().obtenerTipoDocumento(denoTipoDocumento);
-        Distrito distrito = new DistritoFacadeREST().obtenerDistrito(denoDistrito);
-
-        try {
-            em.getTransaction().begin();
-            Datospersonales datosPersonales = em.find(Datospersonales.class, idPersona);
-
-            if (datosPersonales != null) {
                 datosPersonales.setDocuPersona(docuPersona);
                 datosPersonales.setRuc(ruc);
                 datosPersonales.setNombPersona(nombPersona);
@@ -270,16 +239,84 @@ public class DatospersonalesFacadeREST extends AbstractFacade<Datospersonales> {
                 datosPersonales.setIdTipoDocumento(tipoDocumento);
                 datosPersonales.setIdDistrito(distrito);
 
-                em.merge(datosPersonales);
+                em.persist(datosPersonales);
                 em.getTransaction().commit();
                 response.addProperty("success", Boolean.TRUE);
-            } else {
+            } catch (Exception e) {
                 response.addProperty("success", Boolean.FALSE);
             }
-        } catch (Exception e) {
-            response.addProperty("success", Boolean.FALSE);
-        } finally {
-            em.close();
+            response.addProperty("resultado", "ok");
+        } else {
+            response.addProperty("resultado", "error");
+        }
+
+        return gson.toJson(response);
+    }
+
+    // METODO PARA EDITAR LOS DATOS PERSONALES
+    @POST
+    @Path("editardatospersonales")
+    @Consumes(MediaType.APPLICATION_JSON)
+    @Produces(MediaType.APPLICATION_JSON)
+    public String editarDatosPersonales(String data, @Context HttpHeaders headers) throws ParseException {
+        Gson gson = new Gson();
+        JsonObject response = new JsonObject();
+        JsonObject request = JsonParser.parseString(data).getAsJsonObject();
+
+        String tokenClient = extraerTokenHeader(headers);
+        String idToken = JWT.verifyToken(tokenClient);
+
+        if (idToken != null) {
+            int idPersona = request.get("idPersona").getAsInt();
+            String denoTipoDocumento = request.get("denoTipoDocumento").getAsString();
+            String docuPersona = request.get("docuPersona").getAsString();
+            String ruc = request.get("ruc").getAsString();
+            String nombPersona = request.get("nombPersona").getAsString();
+            String apPaPersona = request.get("apPaPersona").getAsString();
+            String apMaPersona = request.get("apMaPersona").getAsString();
+            char genePersona = request.get("genePersona").getAsString().charAt(0);
+            String fecha = request.get("fechPersona").getAsString();
+            String direPersona = request.get("direPersona").getAsString();
+            String celuPersona = request.get("celuPersona").getAsString();
+            String emailPersona = request.get("emailPersona").getAsString();
+            String denoDistrito = request.get("denoDistrito").getAsString();
+
+            SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
+            Date fechPersona = sdf.parse(fecha);
+
+            Tipodocumento tipoDocumento = new TipodocumentoFacadeREST().obtenerTipoDocumento(denoTipoDocumento);
+            Distrito distrito = new DistritoFacadeREST().obtenerDistrito(denoDistrito);
+
+            try {
+                em.getTransaction().begin();
+                Datospersonales datosPersonales = em.find(Datospersonales.class, idPersona);
+
+                if (datosPersonales != null) {
+                    datosPersonales.setDocuPersona(docuPersona);
+                    datosPersonales.setRuc(ruc);
+                    datosPersonales.setNombPersona(nombPersona);
+                    datosPersonales.setApPaPersona(apPaPersona);
+                    datosPersonales.setApMaPersona(apMaPersona);
+                    datosPersonales.setGenePersona(genePersona);
+                    datosPersonales.setFechPersona(fechPersona);
+                    datosPersonales.setDirePersona(direPersona);
+                    datosPersonales.setCeluPersona(celuPersona);
+                    datosPersonales.setEmailPersona(emailPersona);
+                    datosPersonales.setIdTipoDocumento(tipoDocumento);
+                    datosPersonales.setIdDistrito(distrito);
+
+                    em.merge(datosPersonales);
+                    em.getTransaction().commit();
+                    response.addProperty("success", Boolean.TRUE);
+                } else {
+                    response.addProperty("success", Boolean.FALSE);
+                }
+            } catch (Exception e) {
+                response.addProperty("success", Boolean.FALSE);
+            }
+            response.addProperty("resultado", "ok");
+        } else {
+            response.addProperty("resultado", "error");
         }
 
         return gson.toJson(response);
@@ -290,30 +327,47 @@ public class DatospersonalesFacadeREST extends AbstractFacade<Datospersonales> {
     @Path("eliminardatospersonales")
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
-    public String eliminarDatosPersonales(String data) {
+    public String eliminarDatosPersonales(String data, @Context HttpHeaders headers) {
         Gson gson = new Gson();
         JsonObject response = new JsonObject();
         JsonObject request = JsonParser.parseString(data).getAsJsonObject();
 
-        int idPersona = request.get("idPersona").getAsInt();
+        String tokenClient = extraerTokenHeader(headers);
+        String idToken = JWT.verifyToken(tokenClient);
 
-        try {
-            em.getTransaction().begin();
-            Datospersonales datosPersonales = em.find(Datospersonales.class, idPersona);
+        if (idToken != null) {
+            int idPersona = request.get("idPersona").getAsInt();
 
-            if (datosPersonales != null) {
-                em.remove(datosPersonales);
-                em.getTransaction().commit();
-                response.addProperty("success", Boolean.TRUE);
-            } else {
+            try {
+                em.getTransaction().begin();
+                Datospersonales datosPersonales = em.find(Datospersonales.class, idPersona);
+
+                if (datosPersonales != null) {
+                    em.remove(datosPersonales);
+                    em.getTransaction().commit();
+                    response.addProperty("success", Boolean.TRUE);
+                } else {
+                    response.addProperty("success", Boolean.FALSE);
+                }
+            } catch (Exception ex) {
                 response.addProperty("success", Boolean.FALSE);
             }
-            return gson.toJson(response);
-        } catch (Exception ex) {
-            response.addProperty("success", Boolean.FALSE);
-            return gson.toJson(response);
-        } finally {
-            em.close();
+            response.addProperty("resultado", "ok");
+        } else {
+            response.addProperty("resultado", "error");
         }
+
+        return gson.toJson(response);
+    }
+
+    // METODO PARA LA OBTENCION DEL TOKEN CLIENT DE LA AUTHORIZATION BEARER
+    private String extraerTokenHeader(HttpHeaders headers) {
+        String authorizationHeader = headers.getHeaderString(HttpHeaders.AUTHORIZATION);
+
+        if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
+            return authorizationHeader.substring(7);
+        }
+
+        return null;
     }
 }
